+++ config/cypress.webpack.config.js
@@ -5,18 +5,20 @@ const config = require('@redhat-cloud-services/frontend-components-config');
 
 const { config: webpackConfig, plugins } = config({
     rootFolder: resolve(__dirname, '../'),
-    customProxy: [
-        {
-            context: ['/api/inventory/v1/groups'], // you can adjust the `context` value to redirect only specific endpoints
-            target: 'http://localhost:4010', // default prism port
-            secure: false,
-            changeOrigin: true,
-            pathRewrite: { '^/api/inventory/v1': '' },
-            onProxyReq: (proxyReq) => {
-                proxyReq.setHeader('x-rh-identity', 'foobar'); // avoid 401 errors by providing neccessary security header
+    ...process.env.MOCK && {
+        customProxy: [
+            {
+                context: ['/api/inventory/v1/groups'], // you can adjust the `context` value to redirect only specific endpoints
+                target: 'http://localhost:4010', // default prism port
+                secure: false,
+                changeOrigin: true,
+                pathRewrite: { '^/api/inventory/v1': '' },
+                onProxyReq: (proxyReq) => {
+                    proxyReq.setHeader('x-rh-identity', 'foobar'); // avoid 401 errors by providing neccessary security header
+                }
             }
-        }
-    ]
+        ]
+    }
 });
 
 plugins.push(new webpack.DefinePlugin({