+++ src/components/InventoryGroups/utils/api.js
@@ -25,6 +25,21 @@ export const validateGroupName = (name) => {
     .then((resp) => resp?.results.some((group) => group.name === name));
 };
 
+export const getGroupDetail = (groupId) => {
+    return instance.get(`${INVENTORY_API_BASE}/groups/${groupId}`);
+};
+
+export const updateGroupById = (id, payload) => {
+    return instance.patch(`${INVENTORY_API_BASE}/groups/${id}`, {
+        name: payload.name
+    });
+};
+
+export const deleteGroupById = (id) => {
+    return instance.delete(`${INVENTORY_API_BASE}/groups/${id}`);
+
+};
+
 getGroups.propTypes = {
     search: PropTypes.shape({
     // eslint-disable-next-line camelcase
@@ -34,4 +49,8 @@ getGroups.propTypes = {
         perPage: PropTypes.number,
         page: PropTypes.number
     })
+};
+
+getGroupDetail.propTypes = {
+    groupId: PropTypes.string.isRequired
 };